requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bridgexapi
pydantic
